refactor(ci): inline mirror-detection logic into the script
The mirror helpers were used only by the detect script and its test, so
keep them in scripts/detect-unmirrored-images.ts instead of a separate
src module. The executable entry is guarded by import.meta.main so the
exported helpers can be imported by the unit test without side effects.
